+++ src/js/components/Tab/TabContent.tsx
@@ -8,6 +8,7 @@ import ButtonBase from '@material-ui/core/ButtonBase'
 import Typography from '@material-ui/core/Typography'
 import Url from 'components/Tab/Url'
 import { useStore } from 'components/StoreContext'
+import Tooltip from '@material-ui/core/Tooltip'
 
 const pre = `<span style='color:${highlightBorderColor}'>`
 const post = '</span>'
@@ -38,13 +39,29 @@ const useStyles = makeStyles(theme => ({
   },
   duplicated: {
     color: theme.palette.error.light
+  },
+  tooltip: {
+    lineHeight: '1.2rem',
+    userSelect: 'text',
+    whiteSpace: 'normal',
+    wordBreak: 'break-all',
+    wordWrap: 'break-word'
   }
 }))
 
 export default observer(props => {
-  const { userStore } = useStore()
+  const { hoverStore, dragStore, userStore } = useStore()
   const classes = useStyles()
-  const { activate, title, urlCount, focus, isFocused } = props.tab
+  const { faked } = props
+  const {
+    activate,
+    title,
+    url,
+    urlCount,
+    focus,
+    isFocused,
+    isHovered
+  } = props.tab
   const buttonRef = useRef(null)
   const { showUrl, highlightDuplicatedTab } = userStore
   const getHighlightNode = text => {
@@ -61,30 +78,44 @@ export default observer(props => {
     return <div dangerouslySetInnerHTML={{ __html: result.rendered }} />
   }
   useEffect(() => {
-    if (isFocused) {
-      buttonRef.current.focus()
+    const button = buttonRef.current
+    if (!isFocused && document.activeElement === button) {
+      button.blur()
     }
   }, [isFocused])
   const duplicated =
     urlCount > 1 && highlightDuplicatedTab && classes.duplicated
+
+  const { dragging } = dragStore
+  const { hovered } = hoverStore
+  const open = !(faked || dragging || !isHovered || !hovered)
+  const tooltip = (
+    <div className={classes.tooltip}>
+      <p>{title}</p>
+      <p style={{ opacity: 0.8 }}>{url}</p>
+      {urlCount > 1 && <p>There is duplicated tab!</p>}
+    </div>
+  )
   return (
-    <ButtonBase
-      className={classes.ripple}
-      buttonRef={buttonRef}
-      onFocusVisible={focus}
-      onClick={activate}
-      component='div'
-    >
-      <Typography className={classNames(classes.text, duplicated)}>
-        {getHighlightNode(title)}
-      </Typography>
-      {showUrl && (
-        <Url
-          {...props}
-          className={classNames(classes.text, classes.url, duplicated)}
-          getHighlightNode={getHighlightNode}
-        />
-      )}
-    </ButtonBase>
+    <Tooltip {...{ open, title: tooltip }} interactive>
+      <ButtonBase
+        className={classes.ripple}
+        buttonRef={buttonRef}
+        onFocusVisible={focus}
+        onClick={activate}
+        component='div'
+      >
+        <Typography className={classNames(classes.text, duplicated)}>
+          {getHighlightNode(title)}
+        </Typography>
+        {showUrl && (
+          <Url
+            {...props}
+            className={classNames(classes.text, classes.url, duplicated)}
+            getHighlightNode={getHighlightNode}
+          />
+        )}
+      </ButtonBase>
+    </Tooltip>
   )
 })