+++ src/js/components/Window/__tests__/SelectAll.test.tsx
@@ -1,32 +1,22 @@
 import React from 'react'
-import { spy } from 'sinon'
 import { shallow } from 'enzyme'
 import Tooltip from '@material-ui/core/Tooltip'
 import Checkbox from '@material-ui/core/Checkbox'
 import SelectAll from 'components/Window/SelectAll'
-import * as StoreContext from 'components/StoreContext'
 
 const id = 'id'
-const selectAll = spy()
-const unselectAll = spy()
+const toggleSelectAll = jest.fn()
 const props = {
   win: {
     id,
+    toggleSelectAll,
     allTabSelected: true,
     someTabSelected: false,
     disableSelectAll: false
   }
 }
 
-const mockStore = {
-  tabStore: { selectAll, unselectAll }
-}
-
 describe('SelectAll', () => {
-  beforeEach(() => {
-    jest.spyOn(StoreContext, 'useStore').mockImplementation(() => mockStore)
-  })
-
   it('should render correct components', () => {
     const el = shallow(<SelectAll {...props} />)
     expect(el.find(Tooltip).length).toBe(1)
@@ -42,23 +32,13 @@ describe('SelectAll', () => {
     expect(el.find(Tooltip).props().title).toBe('Select all tabs')
   })
 
-  it('should call tabStore.selectAll', () => {
+  it('should call toggleSelectAll when click', () => {
     const el = shallow(<SelectAll {...props} />)
-    const blur = spy()
+    const blur = jest.fn()
     el.find(Checkbox)
       .props()
       .onChange({ target: { blur } })
-    expect(blur.calledOnce).toBe(true)
-    expect(unselectAll.calledOnce).toBe(true)
-  })
-
-  it('should call tabStore.unselectAll', () => {
-    const el = shallow(<SelectAll {...props} win={{ allTabSelected: false }} />)
-    const blur = spy()
-    el.find(Checkbox)
-      .props()
-      .onChange({ target: { blur } })
-    expect(blur.calledOnce).toBe(true)
-    expect(selectAll.calledOnce).toBe(true)
+    expect(blur.mock.calls.length).toBe(1)
+    expect(toggleSelectAll.mock.calls.length).toBe(1)
   })
 })