+++ src/js/libs/colors.tsx
@@ -1,7 +1,9 @@
 import blue from '@material-ui/core/colors/blue'
 import green from '@material-ui/core/colors/green'
 import pink from '@material-ui/core/colors/pink'
+import grey from '@material-ui/core/colors/grey'
 
+export const grey700 = grey[700]
 export const dropTargetColor = green[100]
 export const droppedColor = green[300]
 export const highlightBorderColor = pink.A400